backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
psycopg2-binary
werkzeug
python-dotenv
